2025/2/6
1.使用SO的方式增加了玩家商店，目前有兩個升級狀態:跑速和生命上限。
2.一樣使用SO的方式修改了武器功能升級，修改了武器的傷害增加以及最大彈藥容量上限的功能。
3.更新了場上的玩家商店、武器商店
4.整理了資料夾內容

+++ Assets/Script/物件/武器/GunSetting.cs
@@ -17,7 +17,7 @@ public class GunSetting : MonoBehaviour
     public GameObject fireEffect;
     public float fireDelay;
     public float reloadTime;
-    public float weaponDmg;
+    public int weaponDmg;
     public float spreadRange;
 
     public bool isAuto; //單發或連發
